# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/employees/stats": {
      "get": {
        "operationId": "EmployeesController_getStats",
        "summary": "Employee aggregate stats (no pagination)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Bad Request — Validation failed or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "POST",
                    "error": "Validation failed: email must be a valid email address"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Invalid or expired token"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Missing required permissions: payroll:run"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 500,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/employees/trash": {
      "get": {
        "operationId": "EmployeesController_findTrashed",
        "summary": "List soft-deleted employees",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PRE_BOARDING",
                "ACTIVE",
                "ON_LEAVE",
                "SUSPENDED",
                "EXITED"
              ],
              "type": "string"
            }
          },
          {
            "name": "employmentType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "FULL_TIME",
                "PART_TIME",
                "CONTRACT",
                "INTERN"
              ],
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "designationId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Bad Request — Validation failed or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "POST",
                    "error": "Validation failed: email must be a valid email address"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Invalid or expired token"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Missing required permissions: payroll:run"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 500,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/employees": {
      "post": {
        "operationId": "EmployeesController_create",
        "summary": "Create employee directly (bypasses onboarding)",
        "description": "Validates dept/designation/manager, creates Employee + User in a single transaction with temp password, and sends onboarding invite email.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Bad Request — Validation failed or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "POST",
                    "error": "Validation failed: email must be a valid email address"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Invalid or expired token"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Missing required permissions: payroll:run"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 500,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "EmployeesController_findAll",
        "summary": "List employees",
        "description": "Paginated. Filter by ?status=ACTIVE&departmentId=...&employmentType=FULL_TIME",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PRE_BOARDING",
                "ACTIVE",
                "ON_LEAVE",
                "SUSPENDED",
                "EXITED"
              ],
              "type": "string"
            }
          },
          {
            "name": "employmentType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "FULL_TIME",
                "PART_TIME",
                "CONTRACT",
                "INTERN"
              ],
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "designationId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Bad Request — Validation failed or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "POST",
                    "error": "Validation failed: email must be a valid email address"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Invalid or expired token"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Missing required permissions: payroll:run"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 500,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/employees/{id}/status": {
      "patch": {
        "operationId": "EmployeesController_patchStatus",
        "summary": "Change employee status",
        "description": "Valid transitions: PRE_BOARDING→ACTIVE/EXITED, ACTIVE→ON_LEAVE/SUSPENDED/EXITED. A reason is required when transitioning to SUSPENDED or EXITED. Sending ACTIVE triggers a welcome email.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Employee UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEmployeeStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Bad Request — Validation failed or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "POST",
                    "error": "Validation failed: email must be a valid email address"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Invalid or expired token"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Missing required permissions: payroll:run"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 500,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/employees/{id}/restore": {
      "patch": {
        "operationId": "EmployeesController_restore",
        "summary": "Restore soft-deleted employee",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Employee UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Bad Request — Validation failed or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "POST",
                    "error": "Validation failed: email must be a valid email address"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Invalid or expired token"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Missing required permissions: payroll:run"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 500,
                    "timestamp": "2024-01-15T10:30:00.000Z",
                    "path": "/api/v1/example",
                    "method": "GET",
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "IGreen HRMS API",
    "description": "Enterprise Human Resource Management System API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10
          },
          "departmentId": {
            "type": "string"
          },
          "designationId": {
            "type": "string"
          },
          "payrollStructureId": {
            "type": "string"
          },
          "leavePolicyId": {
            "type": "string"
          },
          "employmentType": {
            "type": "string",
            "enum": [
              "FULL_TIME",
              "PART_TIME",
              "CONTRACT",
              "INTERN"
            ]
          },
          "joiningDate": {
            "type": "string",
            "example": "2026-07-01"
          },
          "reportingManagerId": {
            "type": "string"
          },
          "empCode": {
            "type": "string"
          },
          "probationEndDate": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "pfNumber": {
            "type": "string"
          },
          "esiNumber": {
            "type": "string"
          },
          "uanNumber": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phone",
          "departmentId",
          "designationId",
          "payrollStructureId",
          "leavePolicyId",
          "employmentType",
          "joiningDate"
        ]
      },
...
      "PatchEmployeeStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PRE_BOARDING",
              "ACTIVE",
              "ON_LEAVE",
              "SUSPENDED",
              "EXITED"
            ]
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "description": "Required when transitioning to SUSPENDED or EXITED",
            "example": "Voluntary resignation"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}